Use post render hooks, making everything horribly unsafe in the process
